+++ FlightBookingApp/App.js
@@ -2,34 +2,11 @@ import * as React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {SignUp, Login, MyFlights, Origen, Destino, Fecha, Pasajeros, Confirmacion} from './src/Screens/index';
+import {stackOptions, stackOptionsForm} from './src/Components/StackOptions'
 
 const Stack = createNativeStackNavigator();
 
 const App = () => {
-
-  const stackOptions = {
-    headerStyle: {
-      elevation: 0,
-      shadowOpacity: 0,
-    },
-    headerTintColor: '#5974f5', 
-    headerTitleStyle: {
-      fontSize: 30, 
-      fontWeight: 'bold',
-      left: 0,
-    }, 
-    headerTitleAlign:'left',
-  }
-  const stackOptionsForm = {
-    headerShown: false,
-    headerStyle: {
-      elevation: 0,
-      shadowOpacity: 0,
-    },
-    headerTintColor: '#5974f5', 
-    title: '',
-  }
-
   return (
     <NavigationContainer>
       {
