+++ FlightBookingApp/src/Assets/styled.js
@@ -110,15 +110,11 @@ export const InpL = styled.TextInput`
   padding-left: 3px;
   `;
 
-export const ScrollView = styled.ScrollView`
-
-  `;
-
 export const GoogleBtn = styled.Text`
   text-align: center;
-`
+`;
 
 export const TextAlert = styled.Text`
   font-size: 14px;
   color: red;
-`
+`;
