+++ FlightBookingApp/src/Components/MyFlightComponents/Origen.js
@@ -3,19 +3,10 @@ import React, { useState } from 'react'
 import { Text, View, TouchableOpacity as Button, ToastAndroid } from 'react-native'
 import { Picker } from '@react-native-picker/picker'
 import { styles } from './formStyles'
+import { nextStep, FlightData  } from '../../Assets/hooks/pikerHelper';
+
 //import { useFormHelpers } from './useFormHelpers'
 {/**list picker: https://github.com/react-native-picker/picker */}
-
-const FlightData = {
-    //get the user id. this way i can have it at the end 
-    //of the form when i neeed to upload the form.
-    "userId": "",
-    "Destiny": ["", ""],
-    "Fecha": "", 
-    "Origin": ["", ""], //use setState to get this data, then update the object and pass it
-    "Passengers": "", 
-    "key": 0,
-}
 
 export const Origen = ({navigation}) =>{
     
@@ -23,32 +14,6 @@ export const Origen = ({navigation}) =>{
 
     const [selectedValue, setSelectedValue] = useState("")
     
-    const setObjectValue = (value) =>{
-        //get only the country
-        let Country = value.split(',')
-        Country.shift()
-        Country = Country.join(',')
-
-        setSelectedValue(value)
-        FlightData.Origin[0] = value.slice(0, 3).toUpperCase() //get the first 3 letters of the country
-        FlightData.Origin[1] = Country
-    }
-    
-    const goToScreen = (nextScreen) =>{
-        if (selectedValue != "") {
-            navigation.navigate(nextScreen)
-        }else{
-            ToastAndroid.show("Select a valid option",ToastAndroid.LONG)
-        }
-    }    
-    
-    const nextStep = (pickerValue, nextScreen) =>{
-        setObjectValue(pickerValue)
-        goToScreen(nextScreen)
-    }
-
-    //Next screen Destino
-
     return (
         <View style={styles.screen}>
             {/**Origen */}
@@ -57,7 +22,7 @@ export const Origen = ({navigation}) =>{
                 <View style={(selectedValue != "" ? (styles.input) : (styles.inputDisabled))}>
                     <Picker
                     selectedValue={selectedValue}
-                    onValueChange={(itemValue, itemIndex) => setSelectedValue(itemValue)}
+                    onValueChange={(itemValue, itemIndex) => {setSelectedValue(itemValue)}}
                     >
                         {/**The last ono will be the default value */}
                         <Picker.Item label="Select Location" value=""></Picker.Item> 
@@ -71,7 +36,7 @@ export const Origen = ({navigation}) =>{
                 </View>
             </View>
 
-            <Button style={(selectedValue != "" ? (styles.button) : (styles.buttonDisabled))} onPress={() => nextStep(selectedValue, "Destino")}>
+            <Button style={(selectedValue != "" ? (styles.button) : (styles.buttonDisabled))} onPress={() => nextStep(selectedValue, "Destino", navigation, setSelectedValue, 'O')}>
                 <Text style={styles.centerText}> Next </Text>
             </Button>
