+++ FlightBookingApp/src/Components/StackOptions.js
@@ -1,0 +1,23 @@
+export const stackOptions = {
+    headerStyle: {
+      elevation: 0,
+      shadowOpacity: 0,
+    },
+    headerTintColor: '#5974f5', 
+    headerTitleStyle: {
+      fontSize: 30, 
+      fontWeight: 'bold',
+      left: 0,
+    }, 
+    headerTitleAlign:'left',
+}
+
+export const stackOptionsForm = {
+    headerShown: false,
+    headerStyle: {
+      elevation: 0,
+      shadowOpacity: 0,
+    },
+    headerTintColor: '#5974f5', 
+    title: '',
+}
