--- FlightBookingApp/src/utils/redux/actions.js
@@ -1,8 +1,0 @@
-export const SET_USER_ID = 'SET_USER_ID';
-
-export const setUid = uid => dispatch =>{
-    dispatch({
-        type: SET_USER_ID,
-        payload: uid
-    })
-}
