--- FlightBookingApp/src/utils/redux/store.js
@@ -1,7 +1,0 @@
-import { createStore, combineReducers, applyMiddleware } from 'redux'
-import thunk from 'redux-thunk'
-import { uidReducer } from './reducers.js'
-
-const rootReducer = combineReducers({uidReducer})
-
-export const Store = createStore(rootReducer, applyMiddleware(thunk))
